requirements.txt
flask
flask-cors
torch
transformers
PyPDF2
python-docx
Pillow
pytesseract
openpyxl
